Change type representation in resource dict.
This changes types to be represented as a named tuple in a ResourceDictionary
rather than simply referencing the type directly. This allows handling of
packed resources in cases where the corresponding types are not available or
registered, which is helpful, e.g., in order to determine type dependencies
of a resource.

This also means that ResourceDictionary objects need a resource registry in
order to reconstitute a resource from its dictionary representation. Because of
this, some functionality has to move into resource_registry.